+++ Kvasir/Components/FaceDetection/FaceDetection.h
@@ -11,11 +11,11 @@ class FaceDetection {
 public:
     FaceDetection();
     void DetectFaces(cv::Mat &image, float confidenceLevel, bool display = true);
-    static double CompareFaces(std::vector<double> currentFace, std::vector<double> savedFace, short size = 512);
+    static float CompareFaces(std::vector<float> currentFace, std::vector<float> savedFace, short size = 512);
     size_t GetNumOfFacesDetected();
     cv::dnn::Net GetModel();
 private:
-    static unsigned long numOfFacesDetected;
+    static unsigned int numOfFacesDetected;
     cv::dnn::Net faceDetectionModel;
 };
